Fix vertical layout problems.
There was a gap between the top of the line and the top of the glyphs of the first line that the vertical alignment didn't take into account. Also there was a minor bug in calculating the line height itself that possibly happened when there were glyphs of varying line heights in the line.